refactor(cmake): simplify lab targets and globalize common settings
- Remove redundant output directory overrides in lab directories
- Replace per-target resource copying with a global custom target
- Group common target link libraries into a single interface target
- Moved resources into named folders for `fonts` and `textures`

+++ Lab2/my_drawing.cpp
@@ -241,7 +241,7 @@ void doGraphics(State& gs) {
         drawable_shape.setOrigin({shape_radius, shape_radius});
         drawable_shape.setPosition(shape.position);
 
-        if (i == gs.selected_shape) {
+        if (static_cast<long long>(i) == gs.selected_shape) {
             drawable_shape.setOutlineColor(LIGHT_GRAY);
             drawable_shape.setOutlineThickness(-3.f);
         } else
